Moves alphabet to encrypt class and passes existance test
