+++ lib/encrypt.rb
@@ -1,0 +1,6 @@
+class Encrypt
+
+  def alphabet
+    ("a".."z").to_a << " "
+  end
+end
