+++ lib/enigma.rb
@@ -3,7 +3,5 @@ require_relative '../lib/encrypt'
 
 class Enigma
 
-  def alphabet
-    ("a".."z").to_a << " "
-  end
+  
 end
