+++ test/test_helper.rb
@@ -5,3 +5,4 @@ require 'minitest/pride'
 require 'mocha/minitest'
 require './lib/enigma'
 require './lib/shift'
+require './lib/encrypt'
